PP-2668 Addressing review comments
- added the authorisation test in resource test, which makes sense. We should include one of such test in every resource endpoint
- renamed config variable
