+++ src/test/resources/config/test-it-config.yaml
@@ -61,4 +61,4 @@ jpa:
 baseUrl: ${BASE_URL:-http://localhost:8080}
 productsUiPayUrl: ${PRODUCTSUI_PAY_URL:-http://localhost:3000/pay}
 
-apiToken: ${API_TOKEN:-default-local-api-token}
+productsApiToken: ${PRODUCTS_API_TOKEN:-default-local-api-token}
